1C Connect GUI to BL to Sort Criteria
Implemented the toString methods for both comparators. Implemented
in MainWindow to initialize the sort criteria comboBox so it
stores a TypeComparator and NameComparator. Add a method to trigger
sort criteria when selected that will  sort the ingredients by using
the currently selected criteria. Updated the add and remove methods to
make sure list is sorted after it's been updated.