2C Connect GUI to BL Add Recipe/ingredient
Implemented the method to add selected ingredients to my recipe
list when the ingredeint is selected and button is clicked. Also,
implemented MainWindow to create a Recipe object with the name and list
of ingredeints. Converted the recipe objects into String to be saved
into a file and be formatted like th eutility class expects.