3C Connect UI to BL read/load recipies
Implemented Mainwindow to show recipes from my file that have
the selected ingredient. Also, converted the file into the string
in th text area. Checked if the data loaded and when a new
ingredient was selected the list would load the reciies and remove
the ones that don't include the ingredient.